@@ -1,7 +1,9 @@
-use std::{net::SocketAddr, num::ParseIntError};
+use std::net::SocketAddr;
+use std::num::ParseIntError;
 
-use clap::{clap_derive::ValueEnum, crate_name, Parser};
-use ironrdp_session::{GraphicsConfig, InputConfig};
+use clap::clap_derive::ValueEnum;
+use clap::{crate_name, Parser};
+use ironrdp::session::{GraphicsConfig, InputConfig};
 use sspi::AuthIdentity;
 
 const DEFAULT_WIDTH: u16 = 1920;
